build(front): add platform param

+++ src/front/scripts/removeArtifacts.sh
@@ -1,8 +1,19 @@
 #!/usr/bin/env bash
 set -e
 
+if [ -z "$1" ]; then
+  PLATFORM="all"
+else
+  PLATFORM="$1"
+fi
+
 echo "cleaning up old artifacts"
-rm -f -- *.gz
-rm -f -- *.apk
-rm -f -- *.aab
-rm -f -- *.ipa
+if [ "$PLATFORM" == "all" or "$PLATFORM" == "ios" ]; then
+  rm -f -- *.gz
+  rm -f -- *.ipa
+fi
+
+if [ "$PLATFORM" == "all" or "$PLATFORM" == "android" ]; then
+  rm -f -- *.apk
+  rm -f -- *.aab
+fi
